// client/projects/virkailija-app/src/app/utilities/models/dto/varda-toimipaikka-yhteenveto-dto.model.ts
export class VardaToimipaikkaYhteenvetoDTO {
  vakajarjestaja_nimi: string;
  lapset_lkm: number;
  lapset_maksutieto_voimassaoleva: number;
  lapset_vakapaatos_voimassaoleva: number;
  lapset_vakasuhde_voimassaoleva: number;
  lapset_vuorohoidossa: number;
  lapset_palveluseteli_ja_ostopalvelu: number;
  toimipaikat_voimassaolevat: number;
  toimipaikat_paattyneet: number;
  toimintapainotukset_maara: number;
  kielipainotukset_maara: number;
}